backend/requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==38.0.1
Django==3.2.5
django-cleanup==6.0.0
django-rest-knox==4.2.0
django-timescaledb==0.2.10
djangorestframework==3.12.4
drf-spectacular==0.22.1
gunicorn==20.1.0
idna==3.2
importlib-resources==5.7.1
inflection==0.5.1
jsonschema==4.4.0
psycopg2-binary==2.9.1
pycparser==2.21
pyrsistent==0.18.1
pytz==2021.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.0
